requirements.txt
boto3==1.14.38
pyyaml